requirements.txt
streamlit==1.25.0
tensorflow==2.12.0
Pillow==9.5.0
numpy==1.24.4
scikit-learn==1.3.0
